fix: use numeric casts in reverse SQL and align view name in docs
- RESTORE_M05_VIEW_SQL had float literals (1000000.0, 3600.0) in cost
  arithmetic; replaced with ::numeric casts and integer literals so both
  forward and reverse migration paths keep pricing math in numeric space
- refresh_cost_cells management command help/success text referred to
  current_cost_cells; corrected to pricing_current_cost_cells to match
  the actual DB relation name
- README command comment and management command table updated to the same
  exact name so developers can cross-reference directly with psql

+++ pricing/migrations/0016_expand_current_cost_cells.py
@@ -107,10 +107,10 @@ SELECT
     latest.region,
     latest.hourly_usd,
     latest.scraped_at                               AS pricing_scraped_at,
-    (latest.hourly_usd * bp.tp_size * 1000000.0
-        / (bp.prefill_tps_aggregate * 3600.0))::numeric(10, 4)  AS usd_per_m_input,
-    (latest.hourly_usd * bp.tp_size * 1000000.0
-        / (bp.decode_tps_aggregate  * 3600.0))::numeric(10, 4)  AS usd_per_m_output
+    (latest.hourly_usd * bp.tp_size::numeric * 1000000
+        / (bp.prefill_tps_aggregate::numeric * 3600))::numeric(10, 4)  AS usd_per_m_input,
+    (latest.hourly_usd * bp.tp_size::numeric * 1000000
+        / (bp.decode_tps_aggregate::numeric  * 3600))::numeric(10, 4)  AS usd_per_m_output
 FROM catalog_benchmarkpoint bp
 JOIN latest            ON latest.gpu_id = bp.gpu_id
 JOIN pricing_provider  p ON p.id        = latest.provider_id
